--- crates/builder/op-rbuilder/src/bin/op-rbuilder/main.rs
@@ -1,10 +1,0 @@
-use op_rbuilder::launcher::launch;
-
-// Prefer jemalloc for performance reasons.
-#[cfg(all(feature = "jemalloc", unix))]
-#[global_allocator]
-static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
-
-fn main() -> eyre::Result<()> {
-    launch()
-}
